target-backend/src/checksum.py
import os
import hashlib
import requests
from bs4 import BeautifulSoup
from PIL import Image
import io

def checksum_news(search_keys: list):
    """
    Scrapes news articles based on search keys and calculates a checksum of the content.
    This is a placeholder and does not perform actual scraping.
    """
    all_content = ""
    for key in search_keys:
        # Placeholder for scraping logic
        print(f"Scraping news for: {key}")
        # In a real implementation, you would use requests and BeautifulSoup
        # to get the content from news websites.
        # For this example, we'll just use the key as the content.
        all_content += key

    return hashlib.md5(all_content.encode()).hexdigest()

def checksum_images(search_keys: list):
    """
    Downloads images based on search keys, converts them to grayscale,
    and calculates a checksum of the pixel values.
    This is a placeholder and does not perform actual image downloading/processing.
    """
    total_pixel_sum = 0
    for key in search_keys:
        # Placeholder for image search and download
        print(f"Searching images for: {key}")
        # In a real implementation, you would use a library like google-images-download
        # or a search engine API to get image URLs.
        # For this example, we'll create a dummy image.
        try:
            img = Image.new('RGB', (100, 100), color = 'red')
            grayscale_img = img.convert('L')
            pixel_sum = sum(grayscale_img.getdata())
            total_pixel_sum += pixel_sum
        except Exception as e:
            print(f"Error processing image for key {key}: {e}")

    return hashlib.md5(str(total_pixel_sum).encode()).hexdigest()


def checksum_files(depth: int):
    """
    Calculates a checksum of all file contents from the root directory up to a specified depth.
    """
    all_content = ""
    start_path = '/'
    for root, dirs, files in os.walk(start_path):
        current_depth = root[len(start_path):].count(os.sep)
        if current_depth < depth:
            for name in files:
                try:
                    filepath = os.path.join(root, name)
                    with open(filepath, 'rb') as f:
                        all_content += f.read().hex()
                except (IOError, OSError):
                    # Ignore files that can't be read
                    pass
        elif current_depth >= depth:
             # Prune the directory walk
            del dirs[:]


    return hashlib.md5(all_content.encode()).hexdigest()
