target-backend/requirements.txt
fastapi
sqlalchemy
psycopg2-binary
requests
beautifulsoup4
Pillow
uvicorn
